requirements.txt
discord-py==1.5.1
selenium==3.141.0
tweepy==3.9.0
heroku3==4.2.1

# Google Sheets API packages:
google-auth==1.19.2
google-auth-oauthlib==0.4.1
gspread==3.6.0
oauth2client==4.1.3
oauthlib==3.1.0

# Dropbox:
dropbox==10.3.0